requirements.txt
openai==0.28.1
python-dotenv==1.0.0
streamlit
langchain 